Moved slider editor
