+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/slider/slider.tpl
