+++ lib/assets/core/javascripts/cartodb3/components/form-components/index.js
@@ -31,7 +31,7 @@ require('./editors/list/list-item.js');
 require('./editors/sortable-list.js');
 require('./editors/legend/category-item.js');
 require('./editors/number/number.js');
-require('./editors/slider.js');
+require('./editors/slider/slider.js');
 require('./editors/textarea.js');
 require('./editors/fill.js');
 require('./editors/taglist.js');
